cli refactored
