+++ limix/_cli/__init__.py
@@ -1,0 +1,3 @@
+from ._cli import cli
+
+__all__ = ["cli"]
